+++ src/HaruyasumiRyokouki.Backend/Services/Interfaces/IMediaResolutionCalculationService.cs
@@ -4,7 +4,7 @@ namespace HaruyasumiRyokouki.Backend.Services.Interfaces;
 
 internal interface IMediaResolutionCalculationService
 {
-	int CalculateSize(ImageUrlType linkType, double? dpr, int? resolution);
+	int CalculateSize(ImageUrlType linkType, float? dpr, int? resolution);
 	int GetNearestResolution(int calculatedSize);
-	int GetResolution(ImageUrlType linkType, double? dpr, int? resolution);
+	int GetResolution(ImageUrlType linkType, float? dpr, int? resolution);
 }
